Modernized MTDParametersESModule
-Merged header and source files
-Added consumes calls and used tokens
