+++ Geometry/MTDGeometryBuilder/plugins/MTDParametersESModule.cc
@@ -1,6 +1,7 @@
-#include "MTDParametersESModule.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+#include "FWCore/Framework/interface/ESProducer.h"
 #include "FWCore/Framework/interface/ModuleFactory.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/ESTransientHandle.h"
@@ -10,15 +11,28 @@
 #include "Geometry/MTDGeometryBuilder/interface/MTDParametersFromDD.h"
 #include "CondFormats/GeometryObjects/interface/PMTDParameters.h"
 
-MTDParametersESModule::MTDParametersESModule( const edm::ParameterSet& pset) 
+#include <memory>
+
+class  MTDParametersESModule: public edm::ESProducer
+{
+ public:
+  MTDParametersESModule( const edm::ParameterSet & );
+
+  using ReturnType = std::unique_ptr<PMTDParameters>;
+
+  static void fillDescriptions( edm::ConfigurationDescriptions & );
+  
+  ReturnType produce( const PMTDParametersRcd & );
+
+ private:
+  MTDParametersFromDD builder;
+  const edm::ESGetToken<DDCompactView, IdealGeometryRecord> compactViewToken_;
+ };
+
+MTDParametersESModule::MTDParametersESModule( const edm::ParameterSet& pset) :
+  compactViewToken_{ setWhatProduced(this).consumesFrom<DDCompactView, IdealGeometryRecord>(edm::ESInputTag()) }
 {
   edm::LogInfo("TRACKER") << "MTDParametersESModule::MTDParametersESModule";
-
-  setWhatProduced(this);
-}
-
-MTDParametersESModule::~MTDParametersESModule()
-{ 
 }
 
 void
@@ -32,11 +46,9 @@ MTDParametersESModule::ReturnType
 MTDParametersESModule::produce( const PMTDParametersRcd& iRecord )
 {
   edm::LogInfo("MTDParametersESModule") <<  "MTDParametersESModule::produce(const PMTDParametersRcd& iRecord)" << std::endl;
-  edm::ESTransientHandle<DDCompactView> cpv;
-  iRecord.getRecord<IdealGeometryRecord>().get( cpv );
-    
+  auto cpv = iRecord.getTransientHandle( compactViewToken_ );
   auto ptp = std::make_unique<PMTDParameters>();
-  builder.build( &(*cpv), *ptp );
+  builder.build( cpv.product(), *ptp );
   
   return ptp;
 }
